requirements.txt
flask
flask-cors
pydantic
python-dotenv
pymongo
pytest